feat: Move command to "remote table create", but keep BC
