+++ internal/pkg/service/cli/cmd/remote/table/create/create_table_test.go
@@ -1,4 +1,4 @@
-package table
+package create
 
 import (
 	"fmt"
