+++ internal/pkg/service/cli/helpmsg/msg/remote/table/create/short.txt
